add another screen randomizer source
I tried something new here.  Since I could store values on the zero page, I tried using it as a buffer so that I could quickly spew out a stream of values onto the screen in batches.  I also tried to reuse as many of the subroutines as possible, but it made for too many conditionals, I think.  I just don't know of a way (yet?) to do a fully incremented 16 bit address.  Had to split it into chunks.