Update and rename Random Colors v1 (sequential).txt to Random Colors Displayed in Sequence.txt
This is what I was trying to do in the first place.  Turns out I needed to use a combination of address and register for 16-bit indirect addressing to work.  I should probably reintroduce some of the comments that I just blasted away.  That and label the address variable(s).